# openapi.yaml
openapi: "3.0.3"
info:
  title: Proveny API
  description: |
    Anti-Plagiarism Code Forensics Platform API.
    Establishes proctored authentic baselines and detects AI-assisted code generation
    through sophistication delta analysis, technique genealogy validation, and cohort comparison.
  version: "1.0.0"
  contact:
    name: Proveny Support
    email: support@Proveny.io
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.Proveny.io/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users (Admin only)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortParam"
        - name: role
          in: query
          schema:
            type: string
            enum: [STUDENT, INSTRUCTOR, PROCTOR, ADMIN]
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Users]
      summary: Create a new user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              email: "student@university.edu"
              password: "TempPass456!"
              firstName: "Amir"
              lastName: "Seitkali"
              role: "STUDENT"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user by ID
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags: [Users]
      summary: Update user (Admin only)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # COURSES
  # ============================================================
  /courses:
    get:
      tags: [Courses]
      summary: List courses accessible to the current user
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortParam"
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated list of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCourses"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Courses]
      summary: Create a new course (Instructor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
            example:
              name: "Backend Development 2025"
              code: "CS301-2025"
              weeklyTargets:
                "1": 20
                "2": 25
                "3": 30
                "4": 35
                "5": 40
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /courses/{courseId}/enroll:
    post:
      tags: [Courses]
      summary: Enroll a student in a course (Instructor/Admin)
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId]
              properties:
                studentId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ============================================================
  # SESSIONS (Proctored)
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Obtain via POST /auth/login. Include as: Authorization: Bearer <token>"
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for offset pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
        example: "-createdAt"
      description: "Sort field. Prefix with - for descending. Example: -createdAt, sophisticationScore"
  responses:
    BadRequest:
      description: Bad request — invalid parameters or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 400
            error: "Bad Request"
            message: "sessionCode must be a 6-character alphanumeric string"
    Unauthorized:
      description: Unauthorized — missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 401
            error: "Unauthorized"
            message: "Invalid or expired access token"
    Forbidden:
      description: Forbidden — insufficient role permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 403
            error: "Forbidden"
            message: "Only INSTRUCTOR role can perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 404
            error: "Not Found"
            message: "Submission with id 'abc-123' not found"
    Conflict:
      description: Conflict — resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 409
            error: "Conflict"
            message: "Student already has a baseline for this course"
    UnprocessableEntity:
      description: Unprocessable Entity — validation errors
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            statusCode: 422
            error: "Unprocessable Entity"
            message: "Validation failed"
            details:
              - field: "weekNumber"
                message: "weekNumber must be between 1 and 16"
              - field: "expectedScore"
                message: "expectedScore must be a number between 0 and 100"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            statusCode: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred. Please try again later."
  schemas:
    ErrorResponse:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 87
        totalPages:
          type: integer
          example: 5
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, PROCTOR, ADMIN]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, PROCTOR, ADMIN]
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        isActive:
          type: boolean
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, PROCTOR, ADMIN]
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        instructorId:
          type: string
          format: uuid
        weeklyTargets:
          type: object
          additionalProperties:
            type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateCourseRequest:
      type: object
      required: [name, code, weeklyTargets]
      properties:
        name:
          type: string
          maxLength: 200
        code:
          type: string
          pattern: "^[A-Z]{2,6}[0-9]{3}-[0-9]{4}$"
        weeklyTargets:
          type: object
          additionalProperties:
            type: number
            minimum: 0
            maximum: 100
    PaginatedCourses:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CourseResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        enrolledAt:
          type: string
          format: date-time
...